requirements.txt
streamlit
joblib
openai-whisper
torch
numpy
scikit-learn
ffmpeg-python
